fix: avoid a copy when sending an empty packet
Sending a packet with no payload requires sending only the header. The
code that did that used to chain the header with the payload and then
copy the chained buffers to another buffer. With this change, if the
payload is empty, then it does not even chain the header with the
payload and simply sends the header by itself, as that is already a
contiguous buffer and does not need to be copied to another buffer.